修改解析62指位时，将byte转换为int,否则编译会报错

+++ hall-api/src/main/scala/nirvana/hall/v62/internal/c/gloclib/gaqryqueConverter.scala
@@ -259,7 +259,7 @@ object gaqryqueConverter extends LoggerSupport{
       pstCandData.foreach{ candData =>
         val cand = matchResult.addCandidateResultBuilder()
         cand.setObjectId(candData.szKey)
-        var pos = candData.nIndex
+        var pos = candData.nIndex.toInt
         if(pos > 20){//gafis6.2中平指指位[21,30],所以这里-10
           pos -= 10
         }
